+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [Unreleased changes]
+### Added
+- ArkDoc documentation for the builtins
+
+### Changed
+
+### Removed
+
 ## [3.1.0] - 2021-06-29
 ### Added
 - adding of new string function for manipulation of utf8 string (str:ord and str:chr)