+++ CHANGELOG.md
@@ -6,9 +6,8 @@
 - instructions for `pop` and `pop!` were added, to replace `list:removeAt`
 
 ### Changed
-
+- `list:removeAt` was deprecated
 ### Removed
-- `list:removeAt` is no longer a builtin
 
 ## [3.1.0] - 2021-06-29
 ### Added